separate normalization tests into files
